Restructure crate to resemble typical multi-bin project
